perf(frontend): stop eagerly preloading AgentOverviewPage and recharts
Two more eager-load leaks found by inspecting dist/index.html's
modulepreload list after the route-lazy-loading change:

- useActivityItems.tsx and AgentQuickNav.tsx (both part of the
  always-loaded nav) imported SCENARIOS/getHiddenScenarios/
  useHiddenScenarios directly from AgentOverviewPage.tsx, dragging that
  whole page (PageLayout/PageHeader chain included) into the eager
  bundle despite it being React.lazy()'d in App.tsx. Split that shared
  metadata into scenarioRegistry.tsx so nav code no longer needs to
  import the page module itself.

- Onboarding was kept eager on the (wrong) assumption it's needed
  before routing is known: it's actually a normal post-auth route
  picked at runtime by OnboardingGate, same as every other page. Only
  Login is truly pre-auth. Lazy-loading it removes ConnectAIDialogs
  (~131KB) from the eager set.

- Dropped the manual "recharts-vendor" chunk grouping. recharts/d3 are
  only used by two lazy pages (Dashboard, UserUsage), but forcing them
  into a named vendor chunk made the bundler treat it as always-needed
  and preload it from index.html on every load (~850KB unused outside
  those two routes). Leaving it unnamed lets Rollup fold it into a
  dynamic-import-only shared chunk fetched only when one of those pages
  is actually visited.

Net effect: what a fresh page load must fetch before first paint drops
from ~1.1MB gzip (previously every chunk, since nothing was lazy) to
~430KB gzip.
